rename tool to astbasic
