+++ astinfo/comment.go
@@ -4,7 +4,7 @@ import (
 	"go/ast"
 	"strings"
 
-	"github.com/wanjm/gos/tool"
+	"github.com/wanjm/gos/astbasic"
 )
 
 const TagPrefix = "@gos"
@@ -32,7 +32,7 @@ func parseComment(commentGroup *ast.CommentGroup, commentor Comment) {
 
 // 解析有效的comments
 func parseValidComment(validComment string, commentor Comment) {
-	commands := tool.Fields(validComment) // 多个参数以;分割
+	commands := astbasic.Fields(validComment) // 多个参数以;分割
 	for _, command := range commands {
 		command = strings.Trim(command, " \t")
 		if len(command) == 0 {
