+++ astinfo/field.go
@@ -5,7 +5,7 @@ import (
 	"go/ast"
 	"strings"
 
-	"github.com/wanjm/gos/tool"
+	"github.com/wanjm/gos/astbasic"
 )
 
 type FieldComment struct {
@@ -64,11 +64,11 @@ func (f *FieldBasic) GenVariableCode(goGenerated *GenedFile, wire bool) string {
 
 func (field *Field) parseTag(fieldType *ast.BasicLit) {
 	if fieldType != nil {
-		tag := tool.ConvertGoTag(fieldType.Value)
+		tag := astbasic.ConvertGoTag(fieldType.Value)
 		// if strings.Contains(tag, "wire") {
 		// 	fmt.Print("hello")
 		// }
-		tagList := tool.Fields(tag)
+		tagList := astbasic.Fields(tag)
 		for _, tag := range tagList {
 			kv := strings.Split(tag, ":")
 			if len(kv) == 2 {
