+++ test/tool_test.go
@@ -3,12 +3,12 @@ package test
 import (
 	"testing"
 
-	"github.com/wanjm/gos/tool"
+	"github.com/wanjm/gos/astbasic"
 )
 
 func TestSplit(t *testing.T) {
 	s := "json:\"country\" gorm:\"country\""
-	a := tool.Fields(s)
+	a := astbasic.Fields(s)
 	for _, v := range a {
 		t.Log(v)
 	}
